pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dlcv"
version = "0.0.2"
authors = [
  { name="Arno Nhym", email="arno.nhym@uni-weimar.de" },
]
description = "Second DLCV example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pytest",
    "torchvision",
    "tqdm",
    "matplotlib",
    "seaborn",
    "pycocotools",
    "pyyaml"
]
